requirements.txt
Flask==2.0.2
dnspython==2.1.0
Werkzeug==2.0.2